Move components out of sub folders
