+++ src/components/Alternative.js
@@ -1,5 +1,5 @@
 import React, { PropTypes } from 'react';
-import './Alternative.css';
+import '../css/Alternative.css';
 
 const Alternative = ({ id, text, ...other }) => (
   <div className="Alternative" key={id}>
