+++ src/components/App.js
@@ -1,10 +1,10 @@
 import React from 'react';
-import LiveVoteCount from '../../containers/LiveVoteCount';
-import VoteHandler from '../../containers/VoteHandler';
-import ActiveIssue from '../../containers/ActiveIssue';
-import Heading from '../Heading';
-import ConcludedIssueListContainer from '../../containers/ConcludedIssueListContainer';
-import './App.css';
+import LiveVoteCount from '../containers/LiveVoteCount';
+import VoteHandler from '../containers/VoteHandler';
+import ActiveIssue from '../containers/ActiveIssue';
+import Heading from './Heading';
+import ConcludedIssueListContainer from '../containers/ConcludedIssueListContainer';
+import '../css/App.css';
 
 const App = () => (
   <div className="App">
