+++ src/components/VoteCounter.js
