+++ src/css/ConcludedIssue.css
